Reordered projects and renamed technologies
Technologies -> Skills
Reordered projects
