+++ portfolio/pages/index.js
@@ -2,21 +2,25 @@ import Accomplishments from "../components/Accomplishments/Accomplishments";
 import BgAnimation from "../components/BackgroundAnimation/BackgroundAnimation";
 import Hero from "../components/Hero/Hero";
 import Projects from "../components/Projects/Projects";
-import Technologies from "../components/Technologies/Technologies";
+import Skills from "../components/Skills/Skills";
 import Timeline from "../components/TimeLine/TimeLine";
 import { Layout } from "../layout/Layout";
 import { Section } from "../styles/GlobalComponents";
+import Head from "next/head";
 
 const Home = () => {
   return (
     <Layout>
+      <Head>
+        <title>Portfolio</title>
+      </Head>
       <Section grid>
         <Hero />
         <BgAnimation />
       </Section>
       <Projects />
       <br />
-      <Technologies />
+      <Skills />
       <br />
       <Timeline />
       <Accomplishments />
